+++ src/PluginInfo.cs
@@ -4,6 +4,6 @@
     {
         public const string GUID = "com.kabzon93region.raidadminpanel";
         public const string NAME = "Raid Admin Panel Client";
-        public const string VERSION = "1.0.19";
+        public const string VERSION = "1.0.20";
     }
 }